+++ python_anywhere/registration/admin.py
@@ -1,6 +1,6 @@
 from django.contrib.admin import site
 from django.contrib.auth.admin import UserAdmin
 
-from python_anywhere.accounts.models import User
+from python_anywhere.registration.models import User
 
 site.register(User, UserAdmin)
