+++ python_anywhere/registration/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AccountsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'python_anywhere.accounts'
+    name = 'python_anywhere.registration'
