+++ python_anywhere/registration/templates/registration/signup.html
@@ -11,4 +11,3 @@
         <button type="submit">Sign Up</button>
     </form>
 {% endblock %}
-{#TODO: see if registration can be moved to accounts app, renaming accounts app to registration app#}
