+++ python_anywhere/registration/urls.py
@@ -1,4 +1,4 @@
-# accounts/urls.py
+# registration/urls.py
 from django.urls import path
 
 from .views import SignUpView
