+++ python_anywhere/registration/views.py
@@ -1,9 +1,9 @@
-# accounts/views.py
+# registration/views.py
 from django.contrib.auth.forms import UserCreationForm
 from django.urls import reverse_lazy
 from django.views.generic import CreateView
 
-from accounts.forms import EmailUserCreationForm
+from registration.forms import EmailUserCreationForm
 
 
 class SignUpView(CreateView):
